ENT-4443 Build "brx-spartacus" to dist / Add frontend clean skip opt
The frontend is build into the dist folder. When the module is packaged,
we copy the dist folder contents into the target. The
maven-enforcer-plugin is used to ensure the "dist" folder exists.

By building the frontend into the dist folder, we can still clean the
target folder, and simply build a fresh .jar by copying the contents
of the dist folder again. This is achieved by using the CLI flag
-DskipFE

mvn clean -DskipFE -> this will remove the target folder but not the
"node", "node_modules", "dist", etc. folders.

mvn verify/install -DskipFE -> this will build a new .jar but it will
skip all the NPM/NODE build steps. This requires an existing "dist"
folder.