ENT-4443 Deprecate frontend build in Maven modules
All frontend modules are build by the new Maven "frontend-build" module,
which uses Lerna to install and build. As a temporary step, all the
"frontend-maven-plugin" configuration moved into a "build-legacy"
profile.

To ensure the order in which the Maven modules are build is correct,
all the modules that produces frontend artifacts now depend on
<artifactId>bloomreach-xm-frontend-build</artifactId>.