ENT-6878 Fix vulnerable dependencies in BrxSpartacus
`ts-loader` was completely removed, it should not be a 'dependency' but
a 'devDependency', if anything.

`@xmldom/xmldom` was overridden because we cant update the
`@bloomreach/ng-sdk` because it requires Angular 12. Spartacus 4.x also
requires Angular 12, updating spartacus to a newer version requires a
whole new project so we opted for overriding the problematic version
with the exact patch version that fixes the vulnerability